- pagination and some stats done - can add match - can filter

+++ API/Entities/Photo.cs
@@ -9,6 +9,6 @@ namespace API.Entities
         public string Url { get; set; }
         public string PublicId { get; set; }
         public AppUser AppUser { get; set; }
-        public long AppUserId { get; set; }
+        public int AppUserId { get; set; }
     }
 }
